Move icons to core.
