+++ src/UniversalDashboard.MaterialUI/Components/icon-button.jsx
@@ -4,7 +4,6 @@ import { withStyles } from '@material-ui/core/styles';
 import IconButton from '@material-ui/core/IconButton';
 import classNames from "classnames"
 
-
 const styles = theme => ({
   button: {
     margin: theme.spacing.unit,
