+++ src/UniversalDashboard.Materialize/Components/index.js
@@ -7,7 +7,6 @@ import UDErrorCard from './error-card';
 import UDGrid from './ud-grid';
 import UDLink from './ud-link';
 import UdImageCarousel from './ud-image-carousel';
-import UDIcon from './ud-icon';
 import UDNavbar from './ud-navbar';
 import UDNavigation from './ud-navigation';
 import UdFooter from './ud-footer';
@@ -16,8 +15,6 @@ import UdInputField from './ud-input-field';
 import UdMonitor from './ud-monitor';
 import UDLoading from './ud-loading';
 
-import("font-awesome/css/font-awesome.min.css" /* webpackChunkName: "font-awesome" */);
-
 UniversalDashboard.register("ud-chart", UDChart);
 UniversalDashboard.register("ud-counter", UDCounter);
 UniversalDashboard.register("tab-container", TabContainer);
@@ -25,7 +22,6 @@ UniversalDashboard.register("image-carousel", UdImageCarousel);
 UniversalDashboard.register("ud-errorcard", UDErrorCard);
 UniversalDashboard.register("ud-grid", UDGrid);
 UniversalDashboard.register("ud-footer", UdFooter);
-UniversalDashboard.register("ud-icon", UDIcon);
 UniversalDashboard.register("ud-link", UDLink);
 UniversalDashboard.register("ud-monitor", UdMonitor);
 UniversalDashboard.register("ud-navbar", UDNavbar);
