--- src/UniversalDashboard.Materialize/Components/ud-icon.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default class UdIcon extends React.Component {
-    render() {
-        let iconClass = `fa fa-${this.props.icon}`;
-
-        return <i className={iconClass} style={this.props.style}></i>;
-    }
-}
