+++ src/client/src/app/services/render-service.jsx
@@ -13,12 +13,16 @@ const UdGridComponent = React.lazy(() => import('./../ud-grid.jsx' /* webpackChu
 const UdDateTimeComponent = React.lazy(() => import('./../basics/datetime.jsx' /* webpackChunkName: "ud-date-time" */))
 const UdElementComponent = React.lazy(() => import('./../ud-element.jsx' /* webpackChunkName: "ud-element" */))
 const UdTerminal = React.lazy(() => import( './../ud-terminal.jsx' /* webpackChunkName: "ud-terminal" */))
+const UdIcon = React.lazy(() => import( './../icon.jsx' /* webpackChunkName: "icon" */))
 import UDGridLayout from './../ud-grid-layout';
 
 export function internalRenderComponent(component, history) {
     if (!component) return null;
 
     switch(component.type) {
+
+        case "icon":
+            return <UdIcon {...component} key={component.id}/>;
 
         case "grid-layout":
             return <UDGridLayout {...component} key={component.id}/>;
